refactor: extract padded lines and prompt
This change aims at splitting components that were coupled out of convenience. Specifically, the confirm prompt does not need an input and the search component will never be multiline.

The parts that are in common are really just about PaddedLines, which have now been extracted to their own type and component.

This change also splits components better to improve testability and introduces tests